requirements.txt
fastapi
uvicorn
scikit-learn
pandas
numpy
joblib
openpyxl
imbalanced-learn
google-generativeai
python-dotenv
sqlalchemy
